feat(orchestrator): [scaffold_route] add /diff route, sidebar entry and home card
Scaffold a placeholder Diff Viewer screen wired end-to-end:
- New `/diff` route registered in the router pointing at a placeholder
  `DiffViewer` screen rendering "Coming soon".
- New sidebar entry "Diff Viewer" linking to `/diff`.
- New home FEATURES card "Diff Viewer" navigating to `/diff`.

Refactored sidebar `menuItems` into `app-side-menu.utils.tsx` to
expose `MENU_ITEMS` for unit testing.

Files changed:
- src/screens/diff-viewer/diff-viewer.tsx (new)
- src/screens/diff-viewer/diff-viewer.test.tsx (new)
- src/components/layout/app-side-menu.utils.tsx (new)
- src/components/layout/app-side-menu.utils.test.tsx (new)
- src/components/layout/app-side-menu.tsx (slimmed)
- src/routes/router.tsx (route registered)
- src/screens/home/home.utils.tsx (Diff Viewer feature)
- src/screens/home/home.utils.test.tsx (new)
